+++ lib/Phpfastcache/Drivers/Dynamodb/Driver.php
@@ -18,14 +18,16 @@ namespace Phpfastcache\Drivers\Dynamodb;
 use Aws\Sdk as AwsSdk;
 use Aws\DynamoDb\DynamoDbClient as AwsDynamoDbClient;
 use Aws\DynamoDb\Marshaler as AwsMarshaler;
+use Aws\DynamoDb\Exception\DynamoDbException as AwsDynamoDbException;
 
 use Phpfastcache\Cluster\AggregatablePoolInterface;
-use Phpfastcache\Config\ConfigurationOption;
 use Phpfastcache\Core\Item\ExtendedCacheItemInterface;
 use Phpfastcache\Core\Pool\ExtendedCacheItemPoolInterface;
 use Phpfastcache\Core\Pool\TaggableCacheItemPoolTrait;
 use Phpfastcache\Entities\DriverStatistic;
+use Phpfastcache\Exceptions\PhpfastcacheDriverException;
 use Phpfastcache\Exceptions\PhpfastcacheLogicException;
+use Psr\Http\Message\UriInterface;
 
 /**
  * Class Driver
@@ -36,6 +38,8 @@ use Phpfastcache\Exceptions\PhpfastcacheLogicException;
 class Driver implements ExtendedCacheItemPoolInterface, AggregatablePoolInterface
 {
     use TaggableCacheItemPoolTrait;
+
+    protected const TTL_FIELD_NAME = 't';
 
     protected AwsSdk $awsSdk;
 
@@ -51,6 +55,7 @@ class Driver implements ExtendedCacheItemPoolInterface, AggregatablePoolInterfac
 
     /**
      * @return bool
+     * @throws PhpfastcacheDriverException
      */
     protected function driverConnect(): bool
     {
@@ -63,8 +68,12 @@ class Driver implements ExtendedCacheItemPoolInterface, AggregatablePoolInterfac
         $this->instance = $this->awsSdk->createDynamoDb();
         $this->marshaler = new AwsMarshaler();
 
-        if (!\count($this->instance->listTables(['TableNames' => [$this->getConfig()->getTable()]])->get('TableNames'))) {
+        if (!$this->hasTable()) {
             $this->createTable();
+        }
+
+        if (!$this->hasTtlEnabled()) {
+            $this->enableTtl();
         }
 
         return true;
@@ -78,7 +87,10 @@ class Driver implements ExtendedCacheItemPoolInterface, AggregatablePoolInterfac
     protected function driverWrite(ExtendedCacheItemInterface $item): bool
     {
         $awsItem = $this->marshaler->marshalItem(
-            $this->encodeDocument($this->driverPreWrap($item, true))
+            \array_merge(
+                $this->encodeDocument($this->driverPreWrap($item, true)),
+                ['t' => $item->getExpirationDate()->getTimestamp()]
+            )
         );
 
         $result = $this->instance->putItem([
@@ -136,6 +148,7 @@ class Driver implements ExtendedCacheItemPoolInterface, AggregatablePoolInterfac
 
     /**
      * @return bool
+     * @throws PhpfastcacheDriverException
      */
     protected function driverClear(): bool
     {
@@ -148,8 +161,14 @@ class Driver implements ExtendedCacheItemPoolInterface, AggregatablePoolInterfac
         $this->instance->waitUntil('TableNotExists', $params);
 
         $this->createTable();
+        $this->enableTtl();
 
         return ($result->get('@metadata')['statusCode'] ?? null) === 200;
+    }
+
+    protected function hasTable(): bool
+    {
+        return \count($this->instance->listTables(['TableNames' => [$this->getConfig()->getTable()]])->get('TableNames')) > 0;
     }
 
     protected function createTable() :void
@@ -178,16 +197,69 @@ class Driver implements ExtendedCacheItemPoolInterface, AggregatablePoolInterfac
         $this->instance->waitUntil('TableExists', $params);
     }
 
+    protected function hasTtlEnabled(): bool
+    {
+        $ttlDesc = $this->instance->describeTimeToLive(['TableName' => $this->getConfig()->getTable()])->get('TimeToLiveDescription');
+
+        if (!isset($ttlDesc['AttributeName'], $ttlDesc['TimeToLiveStatus'])) {
+            return false;
+        }
+
+        return $ttlDesc['TimeToLiveStatus'] === 'ENABLED' && $ttlDesc['AttributeName'] === self::TTL_FIELD_NAME;
+    }
+
+    /**
+     * @throws PhpfastcacheDriverException
+     */
+    protected function enableTtl(): void
+    {
+        try {
+            $this->instance->updateTimeToLive([
+                'TableName' => $this->getConfig()->getTable(),
+                'TimeToLiveSpecification' => [
+                    "AttributeName" => self::TTL_FIELD_NAME,
+                    "Enabled" => true
+                ],
+            ]);
+        } catch (AwsDynamoDbException $e) {
+            /**
+             * Error 400 can be an acceptable error of a
+             * Dynamodb restriction: "Time to live has been modified multiple times within a fixed interval"
+             * @see https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_UpdateTimeToLive.html
+             */
+            if ($e->getStatusCode() !== 400) {
+                throw new PhpfastcacheDriverException(
+                    'Failed to enable TTL with the following error: ' . $e->getMessage()
+                );
+            }
+        }
+    }
+
     public function getStats(): DriverStatistic
     {
-        /**
-         * @todo :D
-         */
+        /** @var UriInterface $endpoint */
+        $endpoint = $this->instance->getEndpoint();
+        $table = $this->instance->describeTable(['TableName' => $this->getConfig()->getTable()])->get('Table');
+
+        $info = \sprintf(
+            'Dynamo server "%s" | Table "%s" with %d item(s) stored',
+            $endpoint->getHost(),
+            $table['TableName'] ?? 'Unknown table name',
+            $table['ItemCount'] ?? 'Unknown item count',
+        );
+
+        $data = [
+            'dynamoEndpoint' => $endpoint,
+            'dynamoTable' => $table,
+            'dynamoConfig' => $this->instance->getConfig(),
+            'dynamoApi' => $this->instance->getApi()->toArray(),
+        ];
+
         return (new DriverStatistic())
             ->setData(implode(', ', array_keys($this->itemInstances)))
-            ->setInfo('')
-            ->setRawData([])
-            ->setSize(0);
+            ->setInfo($info)
+            ->setRawData($data)
+            ->setSize($data['dynamoTable']['TableSizeBytes'] ?? 0);
     }
 
     protected function encodeDocument(array $data): array
@@ -204,7 +276,7 @@ class Driver implements ExtendedCacheItemPoolInterface, AggregatablePoolInterfac
         return $data;
     }
 
-    public function getConfig() : Config|ConfigurationOption
+    public function getConfig() : Config
     {
         return $this->config;
     }